pyproject.toml
[project]
name = "fastapi-todo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "python-decouple>=3.8",
]
